chore: move ABConverter to a namespace
