+++ unity-client/Assets/ABConverter/Environment.cs
@@ -1,51 +1,48 @@
 ﻿using System.Collections.Generic;
 
-namespace DCL
+namespace DCL.ABConverter
 {
-    public abstract partial class ABConverter
+    public class Environment
     {
-        public class Environment
+        public readonly IDirectory directory;
+        public readonly IFile file;
+        public readonly IAssetDatabase assetDatabase;
+        public readonly IWebRequest webRequest;
+        public readonly IBuildPipeline buildPipeline;
+
+        public Environment(IDirectory directory, IFile file, IAssetDatabase assetDatabase, IWebRequest webRequest, IBuildPipeline buildPipeline)
         {
-            public readonly IDirectory directory;
-            public readonly IFile file;
-            public readonly IAssetDatabase assetDatabase;
-            public readonly IWebRequest webRequest;
-            public readonly IBuildPipeline buildPipeline;
+            this.directory = directory;
+            this.file = file;
+            this.assetDatabase = assetDatabase;
+            this.webRequest = webRequest;
+            this.buildPipeline = buildPipeline;
+        }
 
-            public Environment(IDirectory directory, IFile file, IAssetDatabase assetDatabase, IWebRequest webRequest, IBuildPipeline buildPipeline)
-            {
-                this.directory = directory;
-                this.file = file;
-                this.assetDatabase = assetDatabase;
-                this.webRequest = webRequest;
-                this.buildPipeline = buildPipeline;
-            }
+        public static Environment CreateWithDefaultImplementations()
+        {
+            return new Environment
+            (
+                directory: new SystemWrappers.Directory(),
+                file: new SystemWrappers.File(),
+                assetDatabase: new UnityEditorWrappers.AssetDatabase(),
+                webRequest: new UnityEditorWrappers.WebRequest(),
+                buildPipeline: new UnityEditorWrappers.BuildPipeline()
+            );
+        }
 
-            public static Environment CreateWithDefaultImplementations()
-            {
-                return new Environment
-                (
-                    directory: new SystemWrappers.Directory(),
-                    file: new SystemWrappers.File(),
-                    assetDatabase: new UnityEditorWrappers.AssetDatabase(),
-                    webRequest: new UnityEditorWrappers.WebRequest(),
-                    buildPipeline: new UnityEditorWrappers.BuildPipeline()
-                );
-            }
+        public static Environment CreateWithMockImplementations()
+        {
+            var file = new Mocked.File();
 
-            public static Environment CreateWithMockImplementations()
-            {
-                var file = new Mocked.File();
-
-                return new Environment
-                (
-                    directory: new Mocked.Directory(),
-                    file: file,
-                    assetDatabase: new Mocked.AssetDatabase(file),
-                    webRequest: new Mocked.WebRequest(),
-                    buildPipeline: new UnityEditorWrappers.BuildPipeline()
-                );
-            }
+            return new Environment
+            (
+                directory: new Mocked.Directory(),
+                file: file,
+                assetDatabase: new Mocked.AssetDatabase(file),
+                webRequest: new Mocked.WebRequest(),
+                buildPipeline: new UnityEditorWrappers.BuildPipeline()
+            );
         }
     }
 }
